treewide: switch "code-oss" to "vscodium"
The "Code - OSS" repository has altered its dependency manager to "npm" which caused numerous issues for nixpkgs.
The most critical issue is the offline cache produced by npm is completely non-reproducible.
As a result, it is currently challenging to update the version of "code-oss" (build-from-source version).

For now, I am switching from "code-oss" to "vscodium" the binary releases maintained by the community.